Use find_packages and move sampling to utils
