+++ prefoil/utils/sampling.py
