+++ tests/test_prefoil.py
@@ -3,7 +3,8 @@ from baseclasses import BaseRegTest
 import numpy as np
 from numpy.testing import assert_allclose, assert_array_equal
 import os
-from prefoil import Airfoil, sampling, generateNACA
+from prefoil import Airfoil, generateNACA
+from prefoil.utils import sampling
 from prefoil.utils import readCoordFile, Error
 from prefoil.utils.geom_ops import _getClosestY
